creates Beer and Store policies, apply policies to Routes, Create, Update and Beer index

+++ app/Livewire/Beers/Index.php
@@ -45,6 +45,7 @@ class Index extends Component
 
     public function remove(Beer $beer): void
     {
+        $this->authorize('delete', $beer);
         $beer->delete();
         Toaster::info("{$beer->name} foi removida com sucesso.");
     }
